[ME-03] style: renamed files from subdirectories_manager* to subdirectories_provider*
